Solution 1: The flexible approach. The Badge tests are stored in an array and could be loaded from somewhere else. That way the function getUsersBadge is only responsible for iterating over the badgeTypes.
Disadvantage: We still have seperate comparison tests for each Badge. This
is quite error prone, since we have to make sure, that the limits for the
intervals line up. One could for example set the upper interval for bronze
to 24 instead of 25. People with exactly 24 solution would not get any
badge. This error could be hard to detect, if unit tests do not cover
edge cases sufficiently.